InfoBubble/Description: consolidate description getting
